Moved lots of stuff around (part 1)
(This is part one of two. Code will not compile between these two commits. This is done to show more easily the big changes versus the ~75 other files that had name refactors)

Major changes, such as file renames and movement of a few thousand lines of Scene.

Scene got a bunch of its contents split into SceneSaver and RenderPreview (to be renamed).

Helper method added to ChunkPosition, with IntBoundingBox class created as return type. This, along with a few other minor changes to code happened around, functionality should be very much the same.